test(skills): pin the middle rewrites, and hold the detector to the repair
TWO RATCHETS, one reported and one found while checking it.

REPORTED. The absolute-zero test asserted only that a balanced document never
GROWS AT THE END. It ignored every middle rewrite, so a regression that invents
new mistaken-closer paths stayed green as long as it did not append.
`FUZZ_MIDDLE_REWRITES` pins the exact count per seed, 3 / 2 / 2, beside the
zero-append assertion. Together they are the whole claim: this many middle
rewrites, which are the tool correctly acting on a mistaken closer, and no
appends at all.

FOUND WHILE MUTATION CHECKING THAT PIN, and the more interesting of the two.
Three separate mutations of the DETECTOR's closing branch moved no repair
behaviour whatsoever. The reason is structural: `find_fence_defects` and
`repair_markdown_fences` are two loops, two hundred lines apart, each with its
own copy of the same grammar. Nothing held them together, so the tool could
report a defect it would not fix, or fix one it would not report.

Reporting a defect and acting on the document are the same question, so they
now have to answer it identically, over every generated document and every
curated case. Measured at introduction: 0 disagreements across three seeds and
136 cases. The detector-only mutation that previously passed every test in the
repository now fails 139.

The additions took the contracts file to 503 lines and the taste ratchet to
576. Raising a baseline is forbidden, so the file splits at the question each
class answers: the contracts file asks whether the scanner agrees with
markdown-it-py, and test_fix_fences_ratchets.py asks whether the scanner's own
pieces agree with each other and with their recorded measurements. 295 and 240
lines; the ratchet is back at 575.

The new module imports the contracts module BY MODULE rather than by name.
Importing the class made pytest collect it twice and the suite reported 1,216
passing where 831 run; a test count that counts the same test twice is the
same species of lie as a stale one.
